requirements.txt
2captcha-python==1.5.1
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
anyio==4.10.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.12.3
billiard==4.2.1
Brotli==1.1.0
capsolver==1.0.7
celery==5.5.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
configparser==6.0.1
contourpy==1.2.1
cryptography==42.0.5
cssselect2==0.8.0
cycler==0.12.1
Django==4.2.16
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
easyocr==1.7.1
etelemetry==0.3.1
exceptiongroup==1.3.1
filelock==3.13.1
fonttools==4.48.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2025.10.0
greenlet==3.2.4
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
imageio==2.33.1
importlib_resources==6.5.2
Jinja2==3.1.3
kiwisolver==1.4.5
kombu==5.5.2
lazy_loader==0.4
looseversion==1.2.0
lxml==4.9.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
ninja==1.13.0
numpy==1.26.4
nvidia-cusparselt-cu12==0.7.1
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.3.0
playwright==1.57.0
prompt_toolkit==3.0.52
propcache==0.4.1
pyclipper==1.3.0.post6
pycparser==2.23
pydyf==0.11.0
pyee==13.0.0
PyJWT==2.10.1
PyMuPDF==1.26.5
pyparsing==3.2.5
PyPDF2==3.0.1
pyphen==0.17.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
PyYAML==6.0.3
qrcode==8.2
reportlab==4.4.6
requests==2.32.5
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.4
sympy==1.14.0
tifffile==2024.8.30
tinycss2==1.4.0
tinyhtml5==2.0.0
torch==1.13.1+cpu
torchvision==0.14.1+cpu
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
yarl==1.22.0
zipp==3.23.0
zopfli==0.2.3.post1
